// src/lib/types/footer.ts
export interface FooterLink {
  label: string;
  href: string;
}

export interface FooterSection {
  title: string;
  links: FooterLink[];
}

export interface SocialLink {
  platform: string;
  href: string;
  icon: string;
}

export interface FooterProps {
  sections: FooterSection[];
  socialLinks: SocialLink[];
  copyright: string;
}